+++ actix-http/CHANGES.md
@@ -5,9 +5,11 @@
 - Minimum supported Rust version (MSRV) is now 1.88.
 - Fix truncated body ending without error when connection closed abnormally. [#3067]
 - Add config/method for `TCP_NODELAY`. [#3918]
+- Do not compress 206 Partial Content responses. [#3191]
 
 [#3067]: https://github.com/actix/actix-web/pull/3067
 [#3918]: https://github.com/actix/actix-web/pull/3918
+[#3191]: https://github.com/actix/actix-web/issues/3191
 
 ## 3.11.2